Move race detector kernel tests to end-to-end
