+++ tests/unit/test_race_detector.py
@@ -1,6 +1,3 @@
-import pytest
-import torch
-
 import triton
 import triton.language as tl
 
@@ -10,24 +7,7 @@ from triton_viz.clients.race_detector.race_detector import (
     SymbolicRaceDetector,
     NullRaceDetector,
 )
-from triton_viz.clients.race_detector.data import AccessEventRecord
 from triton_viz.core.config import config as cfg
-from triton_viz.core.data import Load, Store
-
-
-# ======== Config Isolation ========
-#
-# Kernels are wrapped with @triton_viz.trace inside each test body so the
-# flag-on escape hatch in trace.py (returns the raw kernel when
-# ENABLE_RACE_DETECTOR is off) doesn't strip tracing at import time.
-
-
-@pytest.fixture
-def _isolate_race_detector_cfg():
-    saved = cfg.enable_race_detector
-    cfg.enable_race_detector = True
-    yield
-    cfg.enable_race_detector = saved
 
 
 # ======== Factory Test ========
@@ -45,135 +25,10 @@ def test_race_detector_factory_toggle():
         cfg.enable_race_detector = saved
 
 
-# ======== Raw kernel templates (re-decorated per test) ========
-
-
-@triton.jit
-def _basic_kernel(x_ptr, BLOCK: tl.constexpr):
-    offs = tl.arange(0, BLOCK)
-    v = tl.load(x_ptr + offs)
-    tl.store(x_ptr + offs, v + 1)
-
-
-@triton.jit
-def _load_store_loop_kernel(x_ptr, BLOCK: tl.constexpr):
-    offs = tl.arange(0, BLOCK)
-    for _i in tl.range(0, 4):
-        v = tl.load(x_ptr + offs)
-        tl.store(x_ptr + offs, v + 1)
-
-
-@triton.jit
-def _loop_dedup_kernel(x_ptr, BLOCK: tl.constexpr):
-    offs = tl.arange(0, BLOCK)
-    acc = tl.zeros((BLOCK,), dtype=tl.float32)
-    for _i in tl.range(0, 4):
-        acc += tl.load(x_ptr + offs)
-    tl.store(x_ptr + offs, acc)
-
-
-@triton.jit
-def _loop_premises_kernel(x_ptr, BLOCK: tl.constexpr):
-    offs = tl.arange(0, BLOCK)
-    for _i in tl.range(0, 4):
-        v = tl.load(x_ptr + offs)
-        tl.store(x_ptr + offs, v)
-
-
 @triton.jit
 def _dispatch_kernel(x_ptr, BLOCK: tl.constexpr):
     offs = tl.arange(0, BLOCK)
     tl.store(x_ptr + offs, tl.load(x_ptr + offs))
-
-
-# ======== Basic Capture ========
-
-
-def test_basic_capture(_isolate_race_detector_cfg):
-    detector = SymbolicRaceDetector()
-    traced = triton_viz.trace(client=detector)(_basic_kernel)
-
-    x = torch.zeros(16, dtype=torch.float32)
-    traced[(1,)](x, BLOCK=16)
-
-    modes = [r.access_mode for r in detector.records]
-    op_types = [r.op_type for r in detector.records]
-    assert modes == ["read", "write"], f"expected [read, write], got {modes}"
-    assert op_types == [Load, Store], f"expected [Load, Store], got {op_types}"
-
-    for record in detector.records:
-        assert record.premises, "premises must include addr_ok + pid_ok at minimum"
-        assert record.tensor is not None, "base tensor should be resolvable"
-        assert record.tensor_name == "x_ptr"
-        assert isinstance(record, AccessEventRecord)
-
-
-# ======== Loop: load+store at same address are NOT merged ========
-
-
-def test_loop_load_and_store_same_addr_not_merged(_isolate_race_detector_cfg):
-    detector = SymbolicRaceDetector()
-    traced = triton_viz.trace(client=detector)(_load_store_loop_kernel)
-
-    x = torch.zeros(16, dtype=torch.float32)
-    traced[(1,)](x, BLOCK=16)
-
-    modes = sorted(r.access_mode for r in detector.records)
-    # exactly one read event and one write event (deduped across 4 iterations,
-    # but access_mode splits them so they stay distinct).
-    assert modes == [
-        "read",
-        "write",
-    ], f"load+store at same addr should stay as two events; got {modes}"
-
-
-# ======== Loop: repeated access at same site is deduped ========
-
-
-def test_loop_repeated_access_deduped(_isolate_race_detector_cfg):
-    detector = SymbolicRaceDetector()
-    traced = triton_viz.trace(client=detector)(_loop_dedup_kernel)
-
-    x = torch.zeros(16, dtype=torch.float32)
-    traced[(1,)](x, BLOCK=16)
-
-    reads = [r for r in detector.records if r.access_mode == "read"]
-    assert (
-        len(reads) == 1
-    ), f"loop-body load should dedupe to a single event, got {len(reads)}"
-
-
-# ======== Loop: event carries iterator constraint in its premises ========
-
-
-def test_loop_event_premises_include_iterator(_isolate_race_detector_cfg):
-    detector = SymbolicRaceDetector()
-    traced = triton_viz.trace(client=detector)(_loop_premises_kernel)
-
-    x = torch.zeros(16, dtype=torch.float32)
-    traced[(1,)](x, BLOCK=16)
-
-    assert detector.records
-    for record in detector.records:
-        premise_strs = " ".join(str(p) for p in record.premises)
-        assert "loop_i_" in premise_strs, (
-            f"in-loop event premises must carry a loop iterator constraint; "
-            f"got premises: {premise_strs}"
-        )
-
-
-# ======== String dispatch + ClientManager lookup ========
-
-
-def test_string_dispatch_and_manager_lookup(_isolate_race_detector_cfg):
-    traced = triton_viz.trace("race_detector")(_dispatch_kernel)
-
-    x = torch.zeros(8, dtype=torch.float32)
-    traced[(1,)](x, BLOCK=8)
-
-    rd = traced.client_manager.clients["race_detector"]
-    assert isinstance(rd, SymbolicRaceDetector)
-    assert len(rd.records) == 2
 
 
 # ======== Flag-off escape hatch ========
@@ -193,28 +48,6 @@ def test_flag_off_returns_raw_kernel():
 
         assert not isinstance(traced, TritonTrace)
         assert traced is _dispatch_kernel
-    finally:
-        cfg.enable_race_detector = saved
-
-
-def test_flag_off_does_not_swallow_explicit_instance():
-    """An explicitly constructed detector instance reflects a deliberate user
-    choice and must keep tracing regardless of the global flag."""
-    saved = cfg.enable_race_detector
-    try:
-        cfg.enable_race_detector = False
-        detector = SymbolicRaceDetector()
-        traced = triton_viz.trace(client=detector)(_dispatch_kernel)
-
-        from triton_viz.core.trace import TritonTrace
-
-        assert isinstance(
-            traced, TritonTrace
-        ), "explicit instance must be traced even when the env flag is off"
-
-        x = torch.zeros(8, dtype=torch.float32)
-        traced[(1,)](x, BLOCK=8)
-        assert len(detector.records) == 2
     finally:
         cfg.enable_race_detector = saved
 
@@ -243,35 +76,3 @@ def test_flag_off_returns_raw_kernel_for_factory_instance():
         assert traced is _dispatch_kernel
     finally:
         cfg.enable_race_detector = saved
-
-
-# ======== Repeat launches stay consistent (no partial-grid cache) ========
-
-
-def test_repeat_launches_are_consistent(_isolate_race_detector_cfg):
-    """Launching the same traced kernel twice must produce deterministic,
-    identical block-execution counts. The previous grid-enumeration cache
-    could skip block 0 on a cache hit and then run block 1 instead —
-    different blocks run, different side effects, events captured from a
-    non-deterministic representative. This regression asserts every launch
-    behaves the same."""
-    detector = SymbolicRaceDetector()
-    traced = triton_viz.trace(client=detector)(_dispatch_kernel)
-
-    x1 = torch.zeros(8, dtype=torch.float32)
-    traced[(2,)](x1, BLOCK=8)
-    after_first = list(detector.records)
-
-    detector.records.clear()
-    x2 = torch.zeros(8, dtype=torch.float32)
-    traced[(2,)](x2, BLOCK=8)
-    after_second = list(detector.records)
-
-    assert len(after_first) == len(after_second), (
-        f"repeat launches must capture the same number of events; "
-        f"got {len(after_first)} then {len(after_second)}"
-    )
-    assert [r.access_mode for r in after_first] == [r.access_mode for r in after_second]
-    assert [r.source_location for r in after_first] == [
-        r.source_location for r in after_second
-    ]
